+++ projects/standalone-app/src/app/+feature/state/product/product.facade.ts
@@ -6,7 +6,4 @@ import { Product } from '../../models/product.model';
 
 @Injectable()
 export class ProductFacade extends ProductFacadeBase {
-  constructor(private store: Store<AppState>) {
-    super(Product, store);
-  }
 }