+++ projects/standalone-app/src/app/app.config.ts
@@ -5,11 +5,13 @@ import { routes } from './app.routes';
 import { ConfigService } from './services/config.service';
 
 import { provideAppState } from './state/state.providers';
+import { provideHttpClient } from '@angular/common/http';
 
 
 export const appConfig: ApplicationConfig = {
   providers: [
     provideRouter(routes),
+    provideHttpClient(),
     provideAppState(),
     ConfigService,
   ]