+++ projects/standalone-app/src/app/state/account/account.facade.ts
@@ -3,12 +3,10 @@ import { Store } from '@ngrx/store';
 import { Account } from '../../models/account.model';
 import { AccountFacadeBase, allAccountsForCurrentCustomer } from './account.state';
 import { AppState } from '../app.state';
+import { toSignal } from '@angular/core/rxjs-interop';
 
 @Injectable()
 export class AccountFacade extends AccountFacadeBase {
   readonly currentForCurrentCustomer$ = this.store.select(allAccountsForCurrentCustomer);
-
-  constructor(private store: Store<AppState>) {
-    super(Account, store);
-  }
+  readonly currentForCurrentCustomer = toSignal(this.currentForCurrentCustomer$);
 }