+++ projects/standalone-app/src/app/state/customer/customer.facade.ts
@@ -8,9 +8,5 @@ import { CustomerFacadeBase } from './customer.state';
 
 @Injectable()
 export class CustomerFacade extends CustomerFacadeBase {
-  readonly currentKey = toSignal(this.currentKey$);
-
-  constructor(store: Store<AppState>) {
-    super(Customer, store);
-  }
+  // readonly currentKey = toSignal(this.currentKey$);
 }